merges from 2.2.x: GEOT-895, GEOT-897, PostgisDS.getEnvelope, FIDMapper, GEOT-896
git-svn-id: https://svn.osgeo.org/geotools/geotools/trunk@20666 e5c1c795-43da-0310-a71f-fac65c449510

+++ gt/module/main/src/org/geotools/filter/SQLEncoder.java
@@ -187,8 +187,8 @@ public class SQLEncoder implements org.geotools.filter.FilterVisitor {
         capabilities.addType(FilterCapabilities.NULL_CHECK);
         capabilities.addType(FilterCapabilities.BETWEEN);
         capabilities.addType(FilterCapabilities.FID);
-        capabilities.addType((short) 12345);
-        capabilities.addType((short) -12345);
+        capabilities.addType((long) 12345);
+        capabilities.addType((long) -12345);
 
         return capabilities;
     }
